fix: Compiling issues on msvc

+++ Engine/Code/Source/Rendering/Vulkan/VulkanDescriptor.cpp
@@ -136,7 +136,7 @@ void VulkanDescriptor::UpdateDescriptorSets(IDevice* a_device, ITexture* a_textu
         l_descriptorWrites[1].descriptorCount = 1;
         l_descriptorWrites[1].pImageInfo = &l_imageInfo;
 
-        vkUpdateDescriptorSets(a_device->CastVulkan()->GetDevice(), l_descriptorWrites.size(), l_descriptorWrites.data(), 0, nullptr);
+        vkUpdateDescriptorSets(a_device->CastVulkan()->GetDevice(), static_cast<uint32_t>(l_descriptorWrites.size()), l_descriptorWrites.data(), 0, nullptr);
         DEBUG_LOG_INFO("Vulkan Descriptors : DescriptorSet created!\n");
     }
 }
